Add Storybook Story, Docs Global Template
